requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.10.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
blinker==1.9.0
bootstrap4==0.1.0
certifi==2021.5.30
cffi==2.0.0
charset-normalizer==2.0.4
click==8.3.1
colorama==0.4.4
cors==1.0.1
crispy-tailwind==0.5.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.28
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-ckeditor==6.2.0
django-cors-headers==3.4.0
django-crispy-forms==1.13.0
django-filter==21.1
django-flatpickr==1.0.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
django-notification==1.3.3
django-notifications-hq==1.6.0
django-phonenumber-field==6.0.0
django-simpleui==2025.6.24
django-threadlocals==0.10
django-user-accounts==3.1.0
djangorestframework==3.13.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
hdfs==2.7.3
idna==3.2
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.6
jsonfield==3.1.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Markdown==3.3.6
MarkupSafe==3.0.3
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.2.7
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
py4j==0.10.9.9
pycodestyle==2.7.0
pycparser==2.21
Pygments==2.10.0
PyMySQL==1.1.2
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==4.0.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.26.0
requests-file==3.0.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==2.0.44
sqlparse==0.4.1
swapper==1.3.0
terminado==0.12.1
testpath==0.5.0
tldextract==5.3.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.15.0
tzdata==2025.2
Unipath==1.1
urllib3==1.26.6
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==5.2.0
widgetsnbextension==3.5.2
xlrd==1.2.0
zope.event==6.1
zope.interface==8.1.1
